Remove other component playgrounds
