--- src/components/button/playground.mdx
@@ -1,3 +1,0 @@
-import Button from "./"
-
-<Button label="Hello World" url="/" />
