+++ src/components/dropdown/fixtures.js
@@ -1,0 +1,16 @@
+import React from "react"
+
+// import fixtures from "./menu/fixtures.js"
+// import { DropdownMenu, DropdownTrigger } from "."
+
+export default {
+  simple_dropdown: {
+    children: (
+      <>
+        {/* <DropdownTrigger>Hover over me</DropdownTrigger> */}
+        {/* <DropdownTrigger>Hover over me</DropdownTrigger>
+        <DropdownMenu {...fixtures.simple_menu} /> */}
+      </>
+    )
+  }
+}
