--- src/components/dropdown/menu/__fixtures__/dropdown_menu.json
@@ -1,14 +1,0 @@
-[
-  {
-    "url": "/",
-    "label": "List Item #1"
-  },
-  {
-    "url": "/",
-    "label": "List Item #2"
-  },
-  {
-    "url": "/",
-    "label": "List Item #3"
-  }
-]
