+++ src/components/dropdown/menu/fixtures.js
@@ -1,0 +1,18 @@
+export default {
+  simple_menu: {
+    items: [
+      {
+        url: "/",
+        label: "List Item #1"
+      },
+      {
+        url: "/",
+        label: "List Item #2"
+      },
+      {
+        url: "/",
+        label: "List Item #3"
+      }
+    ]
+  }
+}
