--- src/components/dropdown/playground.mdx
@@ -1,22 +1,0 @@
-import dropdownFixture from "./menu/__fixtures__/dropdown_menu.json"
-import { Dropdown, DropdownMenu, DropdownTrigger } from "./"
-
-# Dropdown
-
-The dropdown component is really more of a set of composable elements that can be used together to build a dropdown menu. It is comprised on three components:
-
-- `<Dropdown />`: A wrapper that controls the visibility of the interior components.
-
-- `<DropdownTrigger />`: A wrapper around a trigger, and doesn't offer any real functional benefit at this time. It lives as a placeholder, should we want to bake functionality into it in the future.
-
-- `<DropdownMenu />`: The element which toggles its visibility based on interaction with the wrapping `<Dropdown />` component.
-
-## Usage
-
-<div>
-  <Dropdown>
-    <DropdownTrigger>Hover over me</DropdownTrigger>
-    <DropdownMenu items={dropdownFixture} />
-  </Dropdown>
-  <p>Here's some other content to demonstrate that the menu is absolutely-positioned by default.</p>
-</div>
