--- src/components/form/playground.mdx
@@ -1,4 +1,0 @@
-import formFixture from "./__fixtures__/form.json"
-import Form from "./"
-
-<Form {...formFixture} />
