+++ src/components/icon/fixtures.js
@@ -1,0 +1,5 @@
+export default {
+  arrow_down: {
+    name: "arrow-down"
+  }
+}
