--- src/components/icon/playground.mdx
@@ -1,13 +1,0 @@
-import Icon from "."
-
-# Icon
-
-The `Icon` component generates an inline svg from `icons.js`, located in `src/components/image/icon`.
-
-- _Prior to adding a new icon to `icons.js` run it through [svgr](https://react-svgr.com/playground/?expandProps=none)._
-
-## Usage
-
-Set the `name` prop as the icon name in `icons.js`.
-
-<Icon name="arrow-down" />
