--- src/components/image/__fixtures__/fixed.js
@@ -1,19 +1,0 @@
-export default {
-  childImageSharp: {
-    fixed: {
-      base64:
-        "data:image/jpeg;base64,/9j/2wBDABALDA4MChAODQ4SERATGCgaGBYWGDEjJR0oOjM9PDkzODdASFxOQERXRTc4UG1RV19iZ2hnPk1xeXBkeFxlZ2P/2wBDARESEhgVGC8aGi9jQjhCY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2NjY2P/wgARCAAGABQDASIAAhEBAxEB/8QAFgABAQEAAAAAAAAAAAAAAAAAAAME/8QAFQEBAQAAAAAAAAAAAAAAAAAAAQL/2gAMAwEAAhADEAAAAbxEyBn/xAAYEAEBAQEBAAAAAAAAAAAAAAACAQMSIv/aAAgBAQABBQLRyx6+WO1//8QAFREBAQAAAAAAAAAAAAAAAAAAARD/2gAIAQMBAT8BJ//EABURAQEAAAAAAAAAAAAAAAAAAAEQ/9oACAECAQE/AWf/xAAZEAEAAgMAAAAAAAAAAAAAAAABABECUWH/2gAIAQEABj8COUzNF3Luf//EABsQAQACAgMAAAAAAAAAAAAAAAEAESExQbHB/9oACAEBAAE/IcGNuDVhkDyRLmodX7P/2gAMAwEAAgADAAAAEIgf/8QAFBEBAAAAAAAAAAAAAAAAAAAAEP/aAAgBAwEBPxAf/8QAFREBAQAAAAAAAAAAAAAAAAAAEEH/2gAIAQIBAT8Qg//EABoQAQEAAwEBAAAAAAAAAAAAAAERACExQbH/2gAIAQEAAT8QB9pFBts9yQRrxRtOuPYIedo+ln//2Q==",
-      width: 1000,
-      height: 286,
-      src:
-        "/static/2eb58d5c70b74dbe16389cd05d535312/a7715/thisisengineering-raeng-b-ilt-cmx-iw-re-unsplash.jpg",
-      srcSet:
-        "/static/2eb58d5c70b74dbe16389cd05d535312/a7715/thisisengineering-raeng-b-ilt-cmx-iw-re-unsplash.jpg 1x",
-      srcWebp:
-        "/static/2eb58d5c70b74dbe16389cd05d535312/e30b5/thisisengineering-raeng-b-ilt-cmx-iw-re-unsplash.webp",
-      srcSetWebp:
-        "/static/2eb58d5c70b74dbe16389cd05d535312/e30b5/thisisengineering-raeng-b-ilt-cmx-iw-re-unsplash.webp 1x",
-      originalName: "thisisengineering-raeng-b-ilt-cmx-iw-re-unsplash.jpg"
-    }
-  }
-}
