--- src/components/link-list/__fixtures__/horizontal.json
@@ -1,17 +1,0 @@
-[
-  { "url": "/", "label": "Vestibulum" },
-  { "url": "#", "label": "Pellentesque venenatis" },
-  { "url": "#", "label": "Fusce" },
-  { "url": "#", "label": "Etiam lacinia dolor" },
-  {
-    "url": "#",
-    "label": "Suspendisse semper",
-    "children": [
-      { "url": "/", "label": "Vestibulum vitae" },
-      { "url": "#", "label": "Pellentesque tempor venenatis" },
-      { "url": "#", "label": "Fusce molestie" },
-      { "url": "#", "label": "Etiam in lacinia dolor" }
-    ]
-  },
-  { "url": "#", "label": "Quisque interdum", "button": true, "className": "button" }
-]
