--- src/components/link-list/__fixtures__/vertical.json
@@ -1,9 +1,0 @@
-{
-  "label": "Quisque",
-  "links": [
-    { "url": "/", "label": "Vestibulum vitae" },
-    { "url": "#", "label": "Pellentesque tempor venenatis" },
-    { "url": "#", "label": "Fusce molestie" },
-    { "url": "#", "label": "Etiam in lacinia dolor" }
-  ]
-}
